+++ sim/semantic/holon_npu_semantic.hpp
@@ -3,9 +3,9 @@
 #include "holon_npu_isa.h"
 #include "holon_npu_program.h"
 #include "holon_npu_runtime.hpp"
+#include "holon_npu_types.hpp"
 
 #include <array>
-#include <compare>
 #include <cstddef>
 #include <cstdint>
 #include <expected>
@@ -75,29 +75,6 @@ using ::holon_npu::runtime::encode_vector_unzip_odd;
 using ::holon_npu::runtime::encode_vector_zip_hi;
 using ::holon_npu::runtime::encode_vector_zip_lo;
 using ::holon_npu::runtime::program_builder;
-
-template <typename Representation, typename Tag>
-class strong_value {
-public:
-    constexpr strong_value() = default;
-    explicit constexpr strong_value(Representation value) : value_(value) {}
-
-    [[nodiscard]] constexpr Representation value() const { return value_; }
-    constexpr auto operator<=>(const strong_value&) const = default;
-
-private:
-    Representation value_{};
-};
-
-struct system_address_tag;
-struct local_address_tag;
-struct instruction_address_tag;
-struct operation_token_tag;
-
-using system_address = strong_value<std::uint64_t, system_address_tag>;
-using local_address = strong_value<std::uint32_t, local_address_tag>;
-using instruction_address = strong_value<std::uint32_t, instruction_address_tag>;
-using operation_token = strong_value<std::uint64_t, operation_token_tag>;
 
 struct decoded_instruction {
     std::uint32_t word = 0;